practice/structlog_demo/requirements.txt
appdirs==1.4.3
cachetools==2.0.0
certifi==2017.4.17
chardet==3.0.3
dill==0.2.6
fluent-logger==0.5.0
future==0.16.0
gapic-google-cloud-logging-v2==0.91.3
git+https://github.com/butla/google-cloud-python@2f24f859a6691a7#egg=google-cloud-logging&subdirectory=logging
googleapis-common-protos==1.5.2
google-auth==1.0.1
google-auth-httplib2==0.0.2
google-cloud-core==0.24.1
google-gax==0.15.13
grpcio==1.3.5
httplib2==0.10.3
idna==2.5
msgpack-python==0.4.8
oauth2client==3.0.0
packaging==16.8
ply==3.8
protobuf==3.3.0
proto-google-cloud-logging-v2==0.91.3
pyasn1==0.2.3
pyasn1-modules==0.0.8
pyparsing==2.2.0
requests==2.16.5
rsa==3.4.2
setuptools==35.0.2
six==1.10.0
structlog==17.2.0
urllib3==1.21.1
